Fix for concat function
Updated the code to accommodate for the CONCAT function or other SQL functions that take more than one argument. Before if a function was used that took more than one argument then wheels would split the argument from inside the function which shouldn't be the case. Fixed that with this by creating the $splitByCommasOutsideFunctions function.
Wrote test cases for this scenario as well.
Wrote one test case that should if $splitByCommasOutsideFunctions function works when table alias is used.
Wrote another test case that should throw an error when table alias are not used.