Removing Evaluates
Removed the usage of the evaluate function and replaced it with the scope mapping to get the scope from the provided value and then used the dynamic allocation operator to get the value from the appropriate scope.
Also used the throw function to throw and error if the value is not found from the request, arguments, application, session and variables scope.