Migration fix and guides update
Updated the wheels.fw in the code to wheels.dev
Updated the guides by fixing mistakes, using wheels-cli instead of cfwheels-cli and updated outdated content.
Added a validation to the migrator.cfm to check if the latestVersion variable exists in the current scope.
Updated the cli.cfm. Removed the include and created an object for the Base.cfc file and used that to call functions.